requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0